Allow script log for specific scripts
