+++ src/main/java/legend/game/debugger/Debugger.java
@@ -6,10 +6,8 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.CheckBox;
 import javafx.scene.control.MenuItem;
 import javafx.stage.Stage;
-import legend.game.Scus94491BpeSegment;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -22,9 +20,6 @@ public class Debugger extends Application {
   private MenuItem menuDebuggersScript;
   @FXML
   private MenuItem menuDebuggersCombat;
-
-  @FXML
-  private CheckBox scriptLog;
 
   public static boolean isRunning() {
     return stage != null;
@@ -54,9 +49,5 @@ public class Debugger extends Application {
   @FXML
   private void showCombatDebugger(final ActionEvent event) throws Exception {
     new CombatDebugger().start(new Stage());
-  }
-
-  public void scriptLogClick(final ActionEvent event) {
-    Scus94491BpeSegment.scriptLog = this.scriptLog.isSelected();
   }
 }
